Move resolveSiteFromUrl tests from site util tests to url tests.
